Point-and-Click Movement (pt. 2)
Implementing changes to both basic movement and point-and-click movement. The player can now click on a non-walkable tile and navigate to its nearest walkable neighbor, if it has one. If not walkable neighbor exists, the player will not move. To support this change, tile click detection now works with the Physics2D.Raycast() method as opposed to the built-in MonoBehaviour.OnMouseDown() method